% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractResults.R
\name{getBestRsCoverages}
\alias{getBestRsCoverages}
\title{Title}
\usage{
getBestRsCoverages(D, rm, tpl, til)
}
\arguments{
\item{D}{D}

\item{rm}{rm}

\item{tpl}{top scores list}

\item{til}{top index matrix list}
}
\value{

}
\description{
Title
}
