% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractResults.R
\name{getCoreK}
\alias{getCoreK}
\title{Determine core K from phase 3 tpl and til}
\usage{
getCoreK(D, rm, tpl, til, cov.thresh, perf.thresh)
}
\arguments{
\item{D}{input matrix D}

\item{rm}{binary rule matrix}

\item{tpl}{list of top performances}

\item{til}{list of top rule set index matrices}

\item{cov.thresh}{core coverage threshold, defaults is 95}

\item{perf.thresh}{core performance threshold, default is 90}
}
\description{
Determine core K from phase 3 tpl and til
}
\examples{
library(crso)
data(skcm)
list2env(skcm.list,envir=globalenv())
Q <- log10(P)
rm.start <- buildRuleLibrary(D,rule.thresh = 0.05) # Rule library matrix, dimension: 60 x 71
til.p2 <- makePhaseTwoImList(D,Q,rm.start,pool.sizes=c(60,10,10),
          max.stored=10,msa=15,shouldPrint = FALSE)
tpl.p2 <- evaluateListOfIMs(D,Q,rm.start,til.p2,msa=15)
core.K <- getCoreK(D,rm.start,tpl.p2,til.p2)
# core.K should be 3 almost always for this example, can run a few time to confirm
}
