% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGeneralizedCore.R
\name{getGCDs}
\alias{getGCDs}
\title{Get Generalized Core Duos}
\usage{
getGCDs(list.subset.cores)
}
\arguments{
\item{list.subset.cores}{list of subset cores}
}
\description{
Get Generalized Core Duos
}
\examples{
list.subset.cores <- list(c("A.B.C","D.E","A.D"),c("A.C","B.C.D","D.E"),
c("A.B.C","D.E"),c("A.B.C","D.E","B.C.D"))
getGCDs(list.subset.cores) # Confidence column should be 100, 100, 100, 75, 50, 25, 25
}
