% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGeneralizedCore.R
\name{getGCEs}
\alias{getGCEs}
\title{Get Generalized Core Events}
\usage{
getGCEs(list.subset.cores)
}
\arguments{
\item{list.subset.cores}{list of subset cores}
}
\description{
Get Generalized Core Events
}
\examples{
list.subset.cores <- list(c("A.B.C","D.E","A.D"),
c("A.C","B.C.D","D.E"),c("A.B.C","D.E"),c("A.B.C","D.E","B.C.D"))
getGCEs(list.subset.cores) # Confidence column should be 100, 100, 100, 100, 100
}
