% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGeneralizedCore.R
\name{getGCRs}
\alias{getGCRs}
\title{Get Generalized Core Rules}
\usage{
getGCRs(list.subset.cores)
}
\arguments{
\item{list.subset.cores}{list of subset cores}
}
\description{
Get Generalized Core Rules
}
\examples{
list.subset.cores <- list(c("A.B.C","D.E","A.D"),c("A.C","B.C.D","D.E"),
c("A.B.C","D.E"),c("A.B.C","D.E","B.C.D"))
getGCRs(list.subset.cores) # Confidence column should be 100, 75, 50, 25, 25
}
