% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_performance_library.R
\name{getSingleRuleWs}
\alias{getSingleRuleWs}
\title{Title}
\usage{
getSingleRuleWs(D, rm, Q, msa)
}
\arguments{
\item{D}{D}

\item{rm}{rm}

\item{Q}{Q}

\item{msa}{msa}
}
\value{

}
\description{
Title
}
