% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase_two_library.R
\name{getTopIm}
\alias{getTopIm}
\title{Title}
\usage{
getTopIm(D, rm, rm.cm, full.im, Q, max.stored, msa, should.check)
}
\arguments{
\item{D}{D}

\item{rm}{rm}

\item{rm.cm}{rm.cm}

\item{full.im}{starting index matrix}

\item{Q}{Q}

\item{max.stored}{max top rules stored}

\item{msa}{msa}

\item{should.check}{Check if valid rule sets? Default is false.}
}
\value{

}
\description{
Title
}
