% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_performance_library.R
\name{getWofAssignment}
\alias{getWofAssignment}
\title{Title}
\usage{
getWofAssignment(D, rm, rs.idc, assign.vec, Q, W.0)
}
\arguments{
\item{D}{binary input D}

\item{rm}{rule library matrix}

\item{rs.idc}{rule set indices}

\item{assign.vec}{assignment vector}

\item{Q}{Q}

\item{W.0}{null penalty}
}
\value{

}
\description{
Title
}
