% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_performance_library.R
\name{getWofRS}
\alias{getWofRS}
\title{Title}
\usage{
getWofRS(D, rm, rm.cm, rs.idc, Q, msa, W.0)
}
\arguments{
\item{D}{D}

\item{rm}{rule library matrix}

\item{rm.cm}{rule coverage matrix}

\item{rs.idc}{rule set indices}

\item{Q}{Q}

\item{msa}{msa}

\item{W.0}{null penalty}
}
\value{

}
\description{
Title
}
