% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_performance_library.R
\name{makeOptimalAssignment}
\alias{makeOptimalAssignment}
\title{Title}
\usage{
makeOptimalAssignment(rm, rm.cm, rs.idc, Q)
}
\arguments{
\item{rm}{rm}

\item{rm.cm}{rm.cm}

\item{rs.idc}{rs.idc}

\item{Q}{penalty matrix}
}
\value{

}
\description{
Title
}
