% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_performance_library.R
\name{makePassengerD}
\alias{makePassengerD}
\title{Title}
\usage{
makePassengerD(D, rm, rs.idc, assign.vec)
}
\arguments{
\item{D}{D}

\item{rm}{rm}

\item{rs.idc}{rs.idc}

\item{assign.vec}{assignment vector}
}
\value{

}
\description{
Title
}
