% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase_four_library.R
\name{makePhaseFourImList}
\alias{makePhaseFourImList}
\title{Title}
\usage{
makePhaseFourImList(
  D,
  Q,
  rm.ordered,
  p3.iml,
  k.max.4,
  max.stored.p4,
  max.nrs.p4,
  msa,
  shouldPrint
)
}
\arguments{
\item{D}{D}

\item{Q}{Q}

\item{rm.ordered}{rm.ordered}

\item{p3.iml}{p3 index matrix list}

\item{k.max.4}{k max 4}

\item{max.stored.p4}{max rules stored per K}

\item{max.nrs.p4}{max rules evaluated}

\item{msa}{minimum samples assigned}

\item{shouldPrint}{Print progress updates? Default is TRUE}
}
\value{

}
\description{
Title
}
