% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase_three_library.R
\name{makePhaseThreeImList}
\alias{makePhaseThreeImList}
\title{Make phase 3 im list from phase 2 im list}
\usage{
makePhaseThreeImList(
  D,
  Q,
  rm.ordered,
  fm.ordered,
  til.ee,
  pool.sizes,
  max.stored,
  max.nrs.p3,
  msa,
  shouldPrint
)
}
\arguments{
\item{D}{binary matrix of events by samples}

\item{Q}{penalty matrix of events by samples}

\item{rm.ordered}{matrix of rules ordered by phase one}

\item{fm.ordered}{family matrix of rm.ordered}

\item{til.ee}{list of rule set matrices (im list) from phase two}

\item{pool.sizes}{pool sizes for phase two}

\item{max.stored}{max number of rule sets saved}

\item{max.nrs.p3}{max number of new rule sets per k, default is 10^5}

\item{msa}{minimum samples assigned per rule in rule set}

\item{shouldPrint}{Print progress updates? Default is TRUE}
}
\value{
phase 3 top im list
}
\description{
Make phase 3 im list from phase 2 im list
}
\examples{
library(crso)
data(skcm)
list2env(skcm.list,envir=globalenv())
Q <- log10(P)
rm.full <- buildRuleLibrary(D,rule.thresh = 0.05) # Rule library matrix, dimension: 60 x 71
til.p2 <- makePhaseTwoImList(D,Q,rm.full,pool.sizes=c(60,10,10),
          max.stored=10,msa=15,shouldPrint = FALSE)
fm.ordered <- getFamMat(rm.full)
til.p3 <- makePhaseThreeImList(D,Q,rm.ordered = rm.full,fm.ordered,
til.ee = til.p2, pool.sizes=c(60,20,20),
         max.stored=100,max.nrs.p3=100,msa=15,shouldPrint = TRUE)
}
