% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{skcm.list}
\alias{skcm.list}
\title{Example data set derived from TCGA skin cutaneous melanoma (SKCM) data.}
\format{
A list with 3 items
\describe{
  \item{D}{Binary alteration matrix. Rows are candidate driver events, columns are samples.}
  \item{P}{Passenger probability matrix corresponding to D.}
  \item{cnv.dictionary}{Data frame containing copy number genes.}
  ...
}
}
\source{
Dataset derived from data generated by the TCGA Research Network: https://www.cancer.gov/tcga
}
\usage{
skcm.list
}
\description{
A dataset containing the processed inputs used in the melanoma analysis within the CRSO publication.
}
\keyword{datasets}
